기존 ViewModel 에서 RoomDatabase 접근 -> useCase 에서 접근
